Move service to pkg
