+++ README.md
@@ -5,7 +5,7 @@
 
 Generates schema reference documentation for Kubernetes Custom Resource Definitions (CRDs).
 
-This tool is built to generate our Management API schema reference in https://docs.giantswarm.io/reference/cp-k8s-api/.
+This tool is built to generate our [Management API CRD schema reference](https://docs.giantswarm.io/ui-api/management-api/crd/).
 
 The generated output consists of Markdown files packed with HTML. By itself, this does not provide a fully readable and user-friendly set of documentation pages. Instead it relies on the HUGO website context, as the [giantswarm/docs](https://github.com/giantswarm/docs) repository, to provide an index page and useful styling.
