+++ pkg/config/testdata/config2.yaml
